Final_Project/requirements.txt
requests==2.32.2
apache-airflow[google]